multiplayer pairing successful

+++ assets/scripts/pair_game.ts
@@ -8,6 +8,7 @@ export default class NewClass extends cc.Component {
 
     private uid: string = null;
     private kick: boolean = false;
+    private key: string = null;
 
     // LIFE-CYCLE CALLBACKS:
 
@@ -16,6 +17,7 @@ export default class NewClass extends cc.Component {
         firebase.auth().onAuthStateChanged((user) => {
             if(user){
                 this.uid = firebase.auth().currentUser.uid;
+                this.key = this.uid.substring(0, 5);
             }else{
                 this.kick = true;
             }
@@ -36,30 +38,61 @@ export default class NewClass extends cc.Component {
             }else this.joinGame()
         }, this);
         cc.find("Canvas/root/back").on(cc.Node.EventType.MOUSE_DOWN, () => {
+            var key: string = this.uid.substring(0, 5);
+            firebase.database().ref('waiting_room/' + key).remove();
+            firebase.database().ref('in_game/' + this.uid).remove();
             this.invite_code.string = '';
             cc.director.loadScene('menu');
         }, this);
     }
 
     makeGame(){
-        var key: string = this.uid.substring(0, 5);
-        this.invite_code.string = key;
-        firebase.database().ref('waiting_room/' + key).set(0);
-        var ref = firebase.database().ref('waiting_room/' + key);
-        ref.on('child_changed', (snapshot) => {
-            var joiner = snapshot.val();
-            ref.remove();
-            firebase.database().ref('in_game/' + joiner + '/creator').set(0);
-            console.log("entering game as creator");
-            // remeber self as creator, then change scene
-            cc.sys.localStorage.setItem("id", 1);
-            cc.sys.localStorage.setItem("room", joiner);
-            cc.director.loadScene('multi');
-        })
+        this.invite_code.string = this.key;
+        // firebase.database().ref('waiting_room/' + this.key).set(0);
+        this.maker_load();
     }
 
-    // update (dt) {}
+    maker_load(){
+        console.log("maker called");
+        firebase.database().ref('waiting_room/' + this.key).once('value', (snapshot) => {
+            console.log(snapshot.val());
+            if(snapshot.exists()){
+                var joiner = snapshot.val();
+                firebase.database().ref('in_game/' + joiner + '/creator').set(0);
+                firebase.database().ref('waiting_room/' + this.key).remove();
+                console.log("entering game as creator");
+                // remeber self as creator, then change scene
+                cc.sys.localStorage.setItem("id", 1);
+                cc.sys.localStorage.setItem("room", joiner);
+                cc.director.loadScene('multi');
+            } else this.maker_load();
+        });
+        // setInterval(this.maker_load, 300);
+    }
+
+    // update (dt) {
+
+    // }
+
     joinGame(){
-        //
+        // var key: string = this.invite_code.string;
+        firebase.database().ref('waiting_room/' + this.invite_code.string).set(this.uid);
+        this.joiner_load();
+    }
+
+    joiner_load(){
+        console.log("joiner called");
+        firebase.database().ref('in_game/' + this.uid).once('value', (snapshot) => {
+            if(snapshot.exists()){
+                console.log("creator has joined");
+                firebase.database().ref('in_game/' + this.uid + '/joiner').set(0);
+                cc.sys.localStorage.setItem("id", 0);
+                cc.sys.localStorage.setItem("room", this.uid);
+                cc.director.loadScene('multi');
+            }else{
+                this.joiner_load();
+            }
+        });
+        // setInterval(this.joiner_load, 300);
     }
 }
